requirements.txt
anthropic>=0.40.0
click>=8.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
